refactor [노은수]: useSWRMutation 활용, 트리거 메서드를 통해, 수동으로 요청 보내도록 수정
